+++ spec/integration/rails/rails_init_spec.rb
@@ -185,8 +185,8 @@ RSpec.describe "plur rails:init command" do
 
         Dir.chdir(dir) do
           result = run_plur("rails:init")
-          expect(result.out).to include("plur db:create")
-          expect(result.out).to include("plur db:migrate")
+          expect(result.out).to include("RAILS_ENV=test plur rails db:create")
+          expect(result.out).to include("RAILS_ENV=test plur rails db:migrate")
           expect(result.out).to include("plur spec")
         end
       end